Revert "Rename: core, next로 디렉토리 정리"
