+++ src/main/java/controller/CreateUserController.java
@@ -1,12 +1,11 @@
-package next.controller;
+package controller;
 
-import core.servlet.AbstractController;
-import next.db.DataBase;
-import core.http.request.HttpRequest;
-import core.http.response.HttpResponse;
-import next.model.User;
+import db.DataBase;
+import http.HttpRequest;
+import http.HttpResponse;
+import model.User;
 
-public class CreateUserController extends AbstractController {
+public class CreateUserController extends AbstractController{
     @Override
     public void doPost(HttpRequest httpRequest, HttpResponse httpResponse) {
         User user = new User(httpRequest.getParameter("userId"), httpRequest.getParameter("password"), httpRequest.getParameter("name"), httpRequest.getParameter("email"));
