+++ src/main/java/db/DataBase.java
@@ -1,11 +1,11 @@
-package next.db;
+package db;
 
 import java.util.Collection;
 import java.util.Map;
 
 import com.google.common.collect.Maps;
 
-import next.model.User;
+import model.User;
 
 public class DataBase {
     private static Map<String, User> users = Maps.newHashMap();
