+++ src/main/java/http/HttpRequest.java
@@ -1,12 +1,9 @@
-package core.http.request;
+package http;
 
-import core.http.HttpCookie;
-import core.http.HttpSession;
-import core.http.HttpSessions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import core.util.HttpRequestUtils;
-import core.util.IOUtils;
+import util.HttpRequestUtils;
+import util.IOUtils;
 
 import java.io.*;
 import java.nio.charset.StandardCharsets;
