+++ src/main/java/http/HttpResponse.java
@@ -1,9 +1,9 @@
-package core.http.response;
+package http;
 
-import core.http.request.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.xml.crypto.Data;
 import java.io.DataOutputStream;
 import java.io.File;
 import java.io.IOException;
