+++ src/main/java/http/HttpSession.java
@@ -1,4 +1,4 @@
-package core.http;
+package http;
 
 import java.util.HashMap;
 import java.util.Map;
