+++ src/main/java/http/HttpSessions.java
@@ -1,6 +1,5 @@
-package core.http;
+package http;
 
-import core.http.request.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
