+++ src/main/java/model/User.java
@@ -1,4 +1,4 @@
-package next.model;
+package model;
 
 public class User {
     private String userId;
