+++ src/main/java/util/HttpRequestUtils.java
@@ -1,4 +1,4 @@
-package core.util;
+package util;
 
 import java.util.Arrays;
 import java.util.Map;
