+++ src/main/java/util/IOUtils.java
@@ -1,4 +1,4 @@
-package core.util;
+package util;
 
 import java.io.BufferedReader;
 import java.io.IOException;
