+++ src/main/java/webserver/RequestHandler.java
@@ -1,14 +1,21 @@
-package core.webserver;
+package webserver;
 
 import java.io.*;
 import java.net.Socket;
+import java.nio.file.Files;
+import java.util.Collection;
+import java.util.Map;
 import java.util.UUID;
 
-import core.servlet.Controller;
-import core.http.request.HttpRequest;
-import core.http.response.HttpResponse;
+import controller.Controller;
+import db.DataBase;
+import http.HttpCookie;
+import http.HttpRequest;
+import http.HttpResponse;
+import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import util.HttpRequestUtils;
 
 public class RequestHandler extends Thread {
     private static final Logger log = LoggerFactory.getLogger(RequestHandler.class);
