+++ src/test/java/http/HttpResponseTest.java
@@ -3,7 +3,6 @@ package http;
 import java.io.*;
 import java.nio.file.Files;
 
-import core.http.response.HttpResponse;
 import org.junit.Test;
 
 public class HttpResponseTest {
